update after 11/17 meeting
- more inline comments
- set the location to northeastern
