+++ studyspaces/DetailSheetView.swift
@@ -1,27 +1,33 @@
 //
-//  LocationDetailsView.swift
+//  DetailSheetView.swift
 //  studyspaces
 //
 //  Created by Josh Cho on 11/12/24.
 //
 
-import SwiftUI
-import MapKit
+// Required frameworks
+import SwiftUI    // For UI components
+import MapKit     // For map and location functionality
 
-struct LocationDetailsView: View {
-    @Binding var mapSelection: MKMapItem?
-    @Binding var show: Bool
-    @State private var lookAroundScene: MKLookAroundScene?
-    @Binding var getDirections: Bool
+// Main view for displaying details about selected location
+struct DetailSheetView: View {
+    // Bindings to communicate with parent view
+    @Binding var mapSelection: MKMapItem?    // Selected location on the map
+    @Binding var show: Bool                  // Controls sheet visibility
+    @State private var lookAroundScene: MKLookAroundScene?    // Stores Apple's street view like preview
+    @Binding var getDirections: Bool         // Triggers navigation
     
     var body: some View {
         VStack {
+            // Header section with location name and close button
             HStack {
                 VStack(alignment: .leading) {
+                    // Location name in large text
                     Text(mapSelection?.placemark.name ?? "")
                         .font(.title2)
                         .fontWeight(.semibold)
                     
+                    // Location address in smaller, gray text
                     Text(mapSelection?.placemark.title ?? "")
                         .font(.footnote)
                         .foregroundStyle(.gray)
@@ -31,6 +37,7 @@ struct LocationDetailsView: View {
                     
                 Spacer()
                     
+                // Close button (X) in top right
                 Button {
                         
                 } label: {
@@ -43,19 +50,24 @@ struct LocationDetailsView: View {
             .padding(.horizontal)
             .padding(.top)
             
+            // Look Around preview section (similar to Google Street View)
             if let scene = lookAroundScene {
+                // Display Look Around preview if available
                 LookAroundPreview(initialScene: scene)
                     .frame(height: 200)
                     .cornerRadius(12)
                     .padding()
             } else {
+                // Show placeholder when preview is unavailable
                 ContentUnavailableView("No preview available", systemImage: "eye.slash")
             }
             
+            // Action buttons container
             HStack(spacing: 24) {
+                // Open in Maps button
                 Button {
                     if let mapSelection {
-                        mapSelection.openInMaps()
+                        mapSelection.openInMaps()    // Opens location in Apple Maps
                     }
                 } label: {
                     Text("Open in Maps")
@@ -66,9 +78,10 @@ struct LocationDetailsView: View {
                         .cornerRadius(12)
                 }
                 
+                // Get Directions button
                 Button {
-                    getDirections = true
-                    show = false
+                    getDirections = true     // Trigger navigation
+                    show = false             // Dismiss sheet
                 } label: {
                     Text("Get Directions")
                         .font(.headline)
@@ -80,10 +93,12 @@ struct LocationDetailsView: View {
             }
             .padding(.horizontal)
         }
+        // Fetch Look Around preview when view appears
         .onAppear {
             print("DEBUG: Did Call on appear")
             fetchLookAroundPreview()
         }
+        // Refresh Look Around preview when selected location changes
         .onChange(of: mapSelection) { oldValue, newValue in
             print("DEBUG: Did call on change")
             fetchLookAroundPreview()
@@ -91,19 +106,21 @@ struct LocationDetailsView: View {
     }
 }
 
-extension LocationDetailsView {
+// MARK: - Helper Methods
+extension DetailSheetView {
+    // Fetches Look Around preview for selected location
     func fetchLookAroundPreview() {
         if let mapSelection {
-            lookAroundScene = nil
+            lookAroundScene = nil    // Clear existing preview
             Task {
                 let request = MKLookAroundSceneRequest(mapItem: mapSelection)
-                lookAroundScene = try? await request.scene
+                lookAroundScene = try? await request.scene    // Fetch new preview asynchronously
             }
         }
     }
 }
 
-
+// Preview provider for SwiftUI canvas
 #Preview {
-    LocationDetailsView(mapSelection: .constant(nil), show: .constant(false), getDirections: .constant(false))
+    DetailSheetView(mapSelection: .constant(nil), show: .constant(false), getDirections: .constant(false))
 }
