+++ studyspaces/studyspacesApp.swift
@@ -11,8 +11,7 @@ import SwiftUI
 struct studyspacesApp: App {
     var body: some Scene {
         WindowGroup {
-//            Text("hello world")
-            SwiftUIView()
+            MapView()
         }
     }
 }
